Only legal moves enabled and Board refactored with Typescript
Added a function for Board to determine if a square is legal to have a piece placed on it. Square
function will now disable it's button if it's not legal to make a move on it.

Board was mistakenly saved as a .jsx file, so it was reformatted into a .tsx file. This required
changing how the for loops iterate amongst other minor tweaks.