Added options to change board size and reset game
The new function Settings controls the options for the game. The options are currently row and
column size. React-hook-forms were used for the forms, since the library offers great performance
and easy testing.

Materialize UI style choices are now set in `Styles.tsx`. This includes the game symbols, which are
now exported from Styles.

Interface types that are shared in multiple files are now set in `Interfaces.tsx`.

generateSquares() was created to move logic outside of the Game() function. Since the `dimensions`
useEffect() updates the squares state, it is no longer necessary to to generate the full board in
initilization.

https://react-hook-form.com/