Refactored file layout
The components were moved into a seperate component folder. Child components that are tightly bound
exist as a sub-component folder for the parent component, but that is the limit to avoid nesting
hell. Each component now has an index.tsx file to serve as a barrel file. This will also aid in
testing when it is implemented.

Styles was moved to its own folder. In the future, specific styles will be included in the component
folder with their respective component.

Files were renamed lowercase to avoid issues with other operating systems.

Interfaces was renamed to types and given it's own folder to serve as the location for custom TS
datatypes that are shared.