Changed component functions type to `function`
This allows the `export default` to be attached on the same line as the function's definition,
making the code easier to read.

The downside of removing `const` is the function can be overwritten; however with component
functions this is unlikely to occur since the whole file is built around them.