Added Dimension type
To clean up multiple different types, the form type was put into a single
Dimensions type.

Prettier was removed since we know use it externally in vscode.

Tests were refactored to use the react-testing-library for consitency with future
tests.

Unused hook functions were removed. Tests that relied on these will be
refactored in the future; for now they are commented out.